+++ test/e2e/specs/NonESM/sourceSwitcherPage.spec.ts
@@ -1,0 +1,8 @@
+import { vpTest } from '../../fixtures/vpTest';
+import { getLinkByName } from '../../testData/pageLinksData';
+import { ExampleLinkName } from '../../testData/ExampleLinkNames';
+
+const link = getLinkByName(ExampleLinkName.SourceSwitcher);
+
+vpTest(`[MOCKED - TODO] Test if video on source switcher page is playing as expected`, async ({ page, pomPages }) => {
+});